# Proto/filter error handling