--- a/filter/all.go
+++ b/filter/all.go
@@ -6,29 +6,4 @@ import (
 )
 
 // All is the list of all baker filters.
-var All = []baker.FilterDesc{
-	ClauseFilterDesc,
-	ClearFieldsDesc,
-	ConcatenateDesc,
-	CountAndTagDesc,
-	CryptDesc,
-	DedupDesc,
-	ExpandJSONDesc,
-	ExpandListDesc,
-	ExternalMatchDesc,
-	FormatTimeDesc,
-	HashDesc,
-	MetadataLastModifiedDesc,
-	MetadataUrlDesc,
-	NotNullDesc,
-	PartialCloneDesc,
-	RegexMatchDesc,
-	ReplaceFieldsDesc,
-	SetStringFromURLDesc,
-	SliceDesc,
-	StringMatchDesc,
-	TimestampDesc,
-	TimestampRangeDesc,
-	URLEscapeDesc,
-	URLParamDesc,
-}
+var All = []baker.FilterDesc{}