--- a/filter/all.go
+++ b/filter/all.go
@@ -7,6 +7,7 @@ import (
 
 // All is the list of all baker filters.
 var All = []baker.FilterDesc{
+	ExternalMatchDesc,
 	NotNullDesc,
 	ReplaceFieldsDesc,
 	URLEscapeDesc,