--- a/filter/all.go
+++ b/filter/all.go
@@ -8,5 +8,6 @@ import (
 // All is the list of all baker filters.
 var All = []baker.FilterDesc{
 	NotNullDesc,
+	ReplaceFieldsDesc,
 	URLEscapeDesc,
 }