--- a/filter/all.go
+++ b/filter/all.go
@@ -12,4 +12,5 @@ var All = []baker.FilterDesc{
 	RegexMatchDesc,
 	ReplaceFieldsDesc,
 	URLEscapeDesc,
+	URLParamDesc,
 }