--- a/filter/all.go
+++ b/filter/all.go
@@ -7,5 +7,6 @@ import (
 
 // All is the list of all baker filters.
 var All = []baker.FilterDesc{
+	NotNullDesc,
 	URLEscapeDesc,
 }