--- a/filter/all.go
+++ b/filter/all.go
@@ -9,6 +9,7 @@ import (
 var All = []baker.FilterDesc{
 	ExternalMatchDesc,
 	NotNullDesc,
+	RegexMatchDesc,
 	ReplaceFieldsDesc,
 	URLEscapeDesc,
 }