--- a/filter/filtertest/passthrough.go
+++ b/filter/filtertest/passthrough.go
@@ -9,11 +9,11 @@ var PassThroughDesc = baker.FilterDesc{
 	Name:   "PassThrough",
 	New:    newPassThrough,
 	Config: &passThrough{},
-	Help:   "lets all through, useful for test/debug purposes",
+	Help:   "let all through, useful for test/debug purposes",
 }
 
 type passThrough struct{}
 
-func newPassThrough(icfg baker.FilterParams) (baker.Filter, error)     { return &passThrough{}, nil }
-func (f *passThrough) Stats() baker.FilterStats                        { return baker.FilterStats{} }
-func (f *passThrough) Process(l baker.Record, next func(baker.Record)) { next(l) }
+func newPassThrough(icfg baker.FilterParams) (baker.Filter, error) { return &passThrough{}, nil }
+func (f *passThrough) Stats() baker.FilterStats                    { return baker.FilterStats{} }
+func (f *passThrough) Process(l baker.Record) error                { return nil }