--- a/filter/notnull.go
+++ b/filter/notnull.go
@@ -2,7 +2,6 @@ package filter
 
 import (
 	"fmt"
-	"sync/atomic"
 
 	"github.com/AdRoll/baker"
 )
@@ -13,6 +12,8 @@ var NotNullDesc = baker.FilterDesc{
 	New:    NewNotNull,
 	Config: &NotNullConfig{},
 	Help:   "Generates errors on records having null (i.e empty) fields.\n",
+
+	DropOnErrorDefault: true,
 }
 
 // NotNullConfig holds configuration parameters for the NotNull filter.
@@ -22,9 +23,8 @@ type NotNullConfig struct {
 
 // NotNull is a baker filter that discards records having null fields.
 type NotNull struct {
-	numFilteredLines int64
-	cfg              *NotNullConfig
-	fields           []baker.FieldIndex
+	cfg    *NotNullConfig
+	fields []baker.FieldIndex
 }
 
 // NewNotNull creates and configures a new NotNull filter.
@@ -44,16 +44,13 @@ func NewNotNull(cfg baker.FilterParams) (baker.Filter, error) {
 
 // Stats implements baker.Filter.
 func (v *NotNull) Stats() baker.FilterStats {
-	return baker.FilterStats{
-		NumFilteredLines: atomic.LoadInt64(&v.numFilteredLines),
-	}
+	return baker.FilterStats{}
 }
 
 // Process implements baker.Filter.
 func (v *NotNull) Process(l baker.Record) error {
 	for _, field := range v.fields {
 		if l.Get(field) == nil {
-			atomic.AddInt64(&v.numFilteredLines, 1)
 			return baker.ErrGenericFilterError
 		}
 	}