--- a/filter_error_handler/all.go
+++ b/filter_error_handler/all.go
@@ -1,0 +1,12 @@
+// Package filter_error_handler provides error handler components for filters.
+package filter_error_handler
+
+import (
+	"github.com/AdRoll/baker"
+)
+
+// All is the list of all baker filter error handlers.
+var All = []baker.FilterErrorHandlerDesc{
+	ClearFieldsDesc,
+	LogDesc,
+}