… bench_test.go
… examples/filter_errors/main.go
… examples/filter_errors/topo.toml
… examples/filtering/filter.go
… filter/clausefilter.go
… filter/clausefilter_test.go
… filter/clear_fields.go
… filter/clear_fields_test.go
… filter/concatenate.go
… filter/concatenate_test.go
… filter/count_and_tag.go
… filter/count_and_tag_test.go
… filter/crypt.go
… filter/crypt_test.go
… filter/dedup.go
… filter/dedup_test.go
… filter/expand_json.go
… filter/expand_json_test.go
… filter/expand_list.go
… filter/expand_list_test.go
… filter/external_match_test.go
… filter/format_time.go
… filter/format_time_test.go
… filter/hash.go
… filter/hash_test.go
… filter/metadata/cache.go
… filter/metadata/cache_test.go
… filter/metadata_lastmodified.go
… filter/metadata_lastmodified_test.go
… filter/metadata_url.go